:building_construction: Replace old architecture with reworked model
Unfortunately, this commit became somewhat of an anti-atomic one. Due to the huge restructuring required and the time invested, it becomes nearly impossible to split this up in any reasonable way since the old and new code base share very few pieces. DO NOT use this as a reference, this is an absolute anti-pattern. However, given the current stage of the project I hope to get away with it :see_no_evil: